cmd640 IDE driver internal spinlocks for config etc accesses.
This is no better or worse than the cli/sti the cmd640 driver
used to have, but at least it compiles and works in the new
scheme of things.

Perfection can wait. Especially since that probably involves
removing the PCI-related code, and just trusting the native
Linux direct PCI accesses.