[PATCH] oss/maestro3.c - convert cli to spinlocks
This is the first of a series of patches against 2.5.32
converting almost all remaining OSS sound drivers to use
spin_lock_irqsave&co instead of cli() which is nowadays
gone when compiling for SMP.

The more complicated ones were:
dmabuf.c
maestro3.c
ad1848.c

I hope I haven't broken anything - but I think some of the
drivers are simply not SMP safe.

One day OSS and some of the older soundcards are obsolete - so what?