[PATCH] show pci_pool stats in driverfs]
This patch exposes basic allocation statistics for pci pools,
very much like /proc/slabinfo but applying to DMA-consistent
memory.  A file "pools" is created in the driverfs directory
for the relevant pci device when the first pool is created, and
removed when the last pool is destroyed.

Please merge to 2.5.latest.  If it matters, DaveM said it
looks fine.  It produces sane output for all the 2.5.30
USB host controller drivers.