kbuild: Fix copying of shipped files
When using cp to copy the shipped file to its actual name,
permissions would be preserved, particularly the copy would be
read-only when the original was (BitKeeper) read-only, leading
to an error when executing the rule a second time.

So now we use cat, which will generate a writable file.